requirement.txt
 django
 djangorestframework
 pandas
 numpy
 scikit-learn
 